pyproject.toml
[tool.poetry]
name = "dbt-scheduler"
version = "0.0.1"
description = "kube-native dbt scheduler"
authors = ["jx2lee <dev.jaejun.lee.1991@gmail.com>"]
readme = "README.md"
packages = [{include = "dbt_scheduler"}]

[tool.poetry.scripts]
dbt-scheduler = "dbt_scheduler.cli.main:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.12.5"
lockfile = "^0.12.2"
psutil = "^6.0.0"
python-daemon = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
